Update Rust sample to modern practices
