+++ .gitignore
@@ -99,7 +99,6 @@ iOSInjectionProject/
 *.class
 
 # Generated files
-bin/
 gen/
 out/
 #  Uncomment the following line in case you need and you don't have the release build type files in your app
