+++ rs/.env.sample
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+# To use this file, rename it from `.env.sample` to `.env` and fill it in
+#
+# Get your AppID and Playground token from https://portal.ditto.live
+
+export APP_ID=""         # Your App ID
+export SHARED_TOKEN=""   # Your Playground Token
+export COLLECTION="cars" # The Collection to watch
